splitting the class
