+++ src/Stratis.Bitcoin.IntegrationTests.Common/EnvironmentMockUpHelpers/GenerateCoinsFastDateTimeProvider.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using NBitcoin;
+
+using Stratis.Bitcoin.Features.Miner.Staking;
 using Stratis.Bitcoin.Signals;
 using Stratis.Bitcoin.Utilities;
 using Stratis.Bitcoin.Utilities.Extensions;
@@ -57,10 +59,10 @@ namespace Stratis.Bitcoin.IntegrationTests.Common.EnvironmentMockUpHelpers
         }
 
         /// <summary>
-        /// This gets called when the coin stake block gets created in <see cref="Features.Miner.PosMinting"/>.
+        /// This gets called when the coin stake block gets created in <see cref="PosMinting"/>.
         /// This gets called when the transaction's time gets set in <see cref="Features.Miner.PowBlockDefinition"/>.
         /// <para>
-        /// Please see the <see cref="Features.Miner.PosMinting.GenerateBlocksAsync"/> method.
+        /// Please see the <see cref="PosMinting.GenerateBlocksAsync"/> method.
         /// </para>
         /// <para>
         /// Please see the <see cref="Features.Miner.PowBlockDefinition.CreateCoinbase"/> method.
